+++ tools/generator/model/src/main/java/is/codion/tools/generator/model/DomainGeneratorModel.java
@@ -106,6 +106,7 @@ public final class DomainGeneratorModel {
 	private static final String AUDIT_COLUMN_NAMES = "auditColumnNames";
 	private static final String PRIMARY_KEY_COLUMN_SUFFIX = "primaryKeyColumnSuffix";
 	private static final String VIEW_SUFFIX = "viewSuffix";
+	private static final String VIEW_PREFIX = "viewPrefix";
 	private static final String LOWER_CASE_IDENTIFIERS = "lowerCaseIdentifiers";
 
 	private final FilterTableModel<SchemaRow, String> schemaTableModel =
@@ -396,6 +397,7 @@ public final class DomainGeneratorModel {
 		json.put(AUDIT_COLUMN_NAMES, schemaSettings.auditColumnNames().stream().collect(joining(",")));
 		json.put(PRIMARY_KEY_COLUMN_SUFFIX, schemaSettings.primaryKeyColumnSuffix());
 		json.put(VIEW_SUFFIX, schemaSettings.viewSuffix());
+		json.put(VIEW_PREFIX, schemaSettings.viewPrefix());
 		json.put(LOWER_CASE_IDENTIFIERS, schemaSettings.lowerCaseIdentifiers());
 
 		PREFERENCES.put(database.name() + "." + tableSchem, json.toString());
@@ -479,6 +481,7 @@ public final class DomainGeneratorModel {
 							.auditColumnNames(json.has(AUDIT_COLUMN_NAMES) ? json.getString(AUDIT_COLUMN_NAMES).split(",") : new String[0])
 							.primaryKeyColumnSuffix(json.optString(PRIMARY_KEY_COLUMN_SUFFIX, null))
 							.viewSuffix(json.optString(VIEW_SUFFIX, null))
+							.viewPrefix(json.optString(VIEW_PREFIX, null))
 							.lowerCaseIdentifiers(json.optBoolean(LOWER_CASE_IDENTIFIERS, true))
 							.build();
 		}